Make getOmedaId its own util for use else where.
This will be use in other places like replacing: https://github.com/parameter1/randall-reilly-websites/pull/709/files#diff-d669545e6e277ea95825e08d608270c5aee5da1edf0c0fd5c6713d16bd4642f5
